Optimizes the DittoLazyProvider for concurrent instance creation.
The previous implementation could cause multiple ditto instances belonging to the same path to be created if the load function was called multiple times before the setDittoHash render cycle kicked in. We now maintain a ref with all the paths that have been created, in order to immediately track the path for which a Ditto instance has been created or is in the process of being created.

The SDK dependency has been updated, and some tests have been optinised to make them more robust.